Add tx to txFilter in handleTX() when needed
This commit ensures that in case the pool has no mempool (e.g. if the
pool belongs to an spv node) the txFilter is updated when the pool
handles a new transaction. This bug was not previously visible because
hasTX() would inadvertently add the transaction to the filter, but after
fixing that bug, new transactions were not included in the txFliter of
pools without a mempool.